Scraping data from NHL.com for database
This application is in progress. It is designed to take data tables from
game sheets published on NHL.com for each game, and pull each event from
these tables. The code will parse and analyze each event, then publish a
database with the relevant information so that it can be manipulated and
analyzed.